requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.8.2
pydantic-settings>=2.3.4
python-dotenv>=1.0.1
httpx>=0.27.0
beautifulsoup4>=4.12.3
lxml>=5.2.2
readability-lxml>=0.8.1
langchain>=0.2.12
langchain-community>=0.2.12
langchain-openai>=0.1.25
tenacity>=8.3.0
loguru>=0.7.2
orjson>=3.10.7
google-api-python-client>=2.100.0
